Moved index code into NavigationServer
Navigation Client and Server are isomorphic files, easier to understand
and document. Makes getProps private. Both client and server are all
private (apart from start in client - matches createServer in server)
