+++ NavigationReact/sample/isomorphic/README.md
@@ -5,6 +5,6 @@ Navigation works just as well on the server as it does on the client. Use it wit
 Once you've cloned the repository, you can install the dependencies and start the isomorphic example:
 
     npm install
-    node index.js
+    node NavigationServer.js
 	
 Then visit http://localhost:8080/ in your browser to see isomorphic Navigation in action.
